Add Lucia auth

+++ src/db/client/tasks.ts
@@ -1,13 +1,21 @@
 import { eq } from 'drizzle-orm'
 import { db } from '@/db'
-import { type Task, type NewTask, tasks, profiles } from '@/db/schema'
+import { type Task, type NewTask, tasks } from '@/db/schema'
+import { isOwner } from '@/db/client/projects'
 
 /**
  * Retrieves a list of tasks for a given project ID.
- * @param forProjectId - The ID of the project to retrieve tasks for.
+ * @param projectId - The ID of the project to retrieve tasks for.
+ * @param userId - The ID of the user making the request.
  * @returns A promise that resolves to an array of tasks.
  */
-export async function listTasksForProject(projectId: string): Promise<Task[]> {
+export async function listTasksForProject(
+  projectId: string,
+  userId: string,
+): Promise<Task[]> {
+  // * Check if the user is an owner of the project
+  if (!(await isOwner(projectId, userId))) return []
+
   return await db.select().from(tasks).where(eq(tasks.projectId, projectId))
 }
 
@@ -15,27 +23,47 @@ export async function listTasksForProject(projectId: string): Promise<Task[]> {
  * Retrieves a list of tasks for the current owner.
  * @returns A promise that resolves to an array of tasks.
  */
-export async function listMyTasks(): Promise<Task[]> {
-  const profileId = 'e444519a-d019-4ca4-bc57-a3263c978d96' /// TODO: Get the current user's profile ID.
-  return await db.select().from(tasks).where(eq(tasks.ownerId, profileId))
+export async function listMyTasks(userId: string): Promise<Task[]> {
+  return await db.select().from(tasks).where(eq(tasks.ownerId, userId))
 }
 
 /**
  * Retrieves a task by its ID.
  * @param id - The ID of the task to retrieve.
+ * @param userId - The ID of the user making the request.
  * @returns A promise that resolves to the task with the specified ID, or null if no task is found.
  */
-export async function getTask(id: string): Promise<Task | null> {
-  const task = await db.select().from(tasks).where(eq(tasks.id, id))
-  return task[0] ?? null
+export async function getTask(
+  id: string,
+  userId: string,
+): Promise<Task | null> {
+  const task = (await db.select().from(tasks).where(eq(tasks.id, id)))[0]
+
+  // * Check if the user is an owner of the project
+  if (!(await isOwner(task.projectId, userId))) return null
+
+  return task ?? null
 }
 
 /**
  * Creates a new task.
  * @param data - The data for the new task.
+ * @param userId - The ID of the user creating the task.
  * @returns A promise that resolves to the created task.
  */
-export async function createTask(data: NewTask): Promise<Task> {
+export async function createTask(
+  data: NewTask,
+  userId: string,
+): Promise<Task | null> {
+  // * Check if the user is an owner of the project
+  if (!(await isOwner(data.projectId, userId))) return null
+
+  // * Check if the owner of the task is an owner of the project
+  if (data.ownerId && !(await isOwner(data.projectId, data.ownerId))) {
+    // * If the owner is not an owner of the project, remove the ownerId from the data
+    delete data.ownerId
+  }
+
   const task = await db.insert(tasks).values(data).returning()
   return task[0]
 }
@@ -44,12 +72,28 @@ export async function createTask(data: NewTask): Promise<Task> {
  * Updates a task with the specified ID.
  * @param id - The ID of the task to update.
  * @param data - The new task data.
+ * @param userId - The ID of the user making the request.
  * @returns A promise that resolves to the updated task.
  */
 export async function updateTask(
   id: string,
   data: Partial<Task>,
-): Promise<Task> {
+  userId: string,
+): Promise<Task | null> {
+  /// Get task to check if the user is an owner of the project
+  const originalTask = await getTask(id, userId)
+
+  // * Check if the user is an owner of the project
+  if (!originalTask) return null
+
+  /// Also make sure that a new owner of the task is an owner of the project
+  if (data.ownerId && data.ownerId !== originalTask.ownerId) {
+    if (!(await isOwner(originalTask.projectId, data.ownerId))) {
+      // * If the new owner is not an owner of the project, remove the ownerId from the data
+      delete data.ownerId
+    }
+  }
+
   const task = await db
     .update(tasks)
     .set(data)
@@ -61,8 +105,15 @@ export async function updateTask(
 /**
  * Deletes a task from the database.
  * @param id - The ID of the task to delete.
+ * @param userId - The ID of the user making the request.
  * @returns A promise that resolves when the task is successfully deleted.
  */
-export async function deleteTask(id: string): Promise<void> {
+export async function deleteTask(id: string, userId: string): Promise<void> {
+  /// Get task to check if the user is an owner of the project
+  const originalTask = await getTask(id, userId)
+
+  // * Check if the user is an owner of the project
+  if (!originalTask) return
+
   await db.delete(tasks).where(eq(tasks.id, id))
 }
